SIMSBIOHUB-668 addressed all items in PR feedback, other small improvements
